Use AM_XGETTEXT_OPTION to add --from-code=UTF-8
Add --from-code=UTF-8 with AM_XGETTEXT_OPTION instead of
po/Makevars.template, so that we can remove the file
po/Makevars.template from the git repo like in all the
libexif repos.

AM_XGETTEXT_OPTION was introduced in gettext 0.17, so the
old requirement for gettext >= 0.14.1 is insufficient.

This updates the gettext requirement to 0.18.3, which is
from 2013-06, which is still earlier than the 2013-12
release date of our other requirement automake 1.14.1, so
this should be fine for most people in 2021.