Add icons & desktop file
The original design is by Valère Monseur & Daniel Fandrich. It was
created in 2013 but it was forgotten to be checked in until now.
gexif-master.svg is the original image that should be edited to change
the icon. The version in scalable/ has had the type replaced by paths so
that the icon can be rendered without special fonts being installed.