Avoid rebuilding icons when unnecessary
The time stamps of the bitmapped and scalable icons depends on the SCCS
and if the svgz file ends up newer than the png files, make will try to
rebuild them unnecessarily. Add a dist-hook to ensure the svgz has an
older time stamp in the tar ball to avoid this. Also, create icon
directories to fix out-of-tree builds